CSS and moved find-games to root of comp folder
